refactor: move Azure Content Safety client to separate namespace
