+++ dotnet/packages/Microsoft.Bot.Builder.M365/Microsoft.Bot.Builder.M365/AI/AzureContentSafety/AzureContentSafetyClientOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.Bot.Builder.M365.AI.OpenAI
+﻿namespace Microsoft.Bot.Builder.M365.AI.AzureContentSafety
 {
     public class AzureContentSafetyClientOptions
     {
